+++ rskj-core/src/main/java/org/ethereum/config/blockchain/mainnet/MainNetOrchidConfig.java
@@ -1,7 +1,6 @@
-package org.ethereum.config.blockchain.testnet;
+package org.ethereum.config.blockchain.mainnet;
 
-// TODO: find a proper name for the "FirstFork"
-public class TestNetFirstForkConfig extends TestNetAfterBridgeSyncConfig {
+public class MainNetOrchidConfig extends MainNetAfterBridgeSyncConfig {
     @Override
     public boolean isRfs50() {
         return true;
